requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
dataclasses-json==0.5.7
filelock==3.9.0
frozenlist==1.3.3
gpt-index==0.2.1
greenlet==2.0.1
huggingface-hub==0.11.1
idna==3.4
joblib==1.2.0
langchain==0.0.59
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==0.4.3
nltk==3.8.1
numpy==1.24.1
openai==0.26.0
packaging==23.0
pandas==1.5.2
pydantic==1.10.4
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.46
tenacity==8.1.0
tokenizers==0.13.2
tqdm==4.64.1
transformers==4.25.1
typing-extensions==4.4.0
typing-inspect==0.8.0
urllib3==1.26.13
yarl==1.8.2
